move chat to cli
